breaking: reimplement dictionary model
